refactor: adapt test modules to structure changes
