+++ setup.cfg
@@ -40,7 +40,10 @@ keywords =
 zip_safe = True
 install_requires =
     depinfo
+    numpy
+    pandas
     cobra
+    altair
 python_requires = >=2.7,!=3.1.*,!=3.2.*,!=3.3.*,!=3.4.*
 packages = find:
 
@@ -71,7 +74,10 @@ known_first_party = driven
 known_third_party =
     depinfo
     pytest
+    numpy
+    pandas
     cobra
+    altair
 
 # See the docstring in versioneer.py for instructions. Note that you must
 # re-run 'versioneer.py setup' after changing this section, and commit the
